Fix portability issues (#2813)
This PR fixes various issues related to portability to 64-bit compilers, including apple clang.

- Header-only libraries.
  ArduinoJsonX, WebCam and flatbuffers do not contain any source code to compile to an archive library.
  The GNU linker is tolerant of this, but not so with other linkers.
  These must set `COMPONENT_LIBNAME=0` so archive isn't created.
- Patch serialllib.
  O_NDELAY is not portable, an alias for O_NONBLOCK which is.
  bzero is obsolete, use memset.
- Fix sockets strerror_r (also used in IFS library, fix in #2224).
  Issue is same function with different return signature.
- Can use `std::gcd` with C++17. Prior to this needed `__gcd` which is not portable.
- Fix unused variable, missing headers, redefined macro warnings
- Fix build errors when compiling 64-bit code.
  Some types are now 64-bit and we get casting or conversion errors.

+++ Sming/Wiring/WVector.h
@@ -182,7 +182,7 @@ public:
 		removeAllElements();
 	}
 
-	bool ensureCapacity(unsigned int minCapacity);
+	bool ensureCapacity(size_t minCapacity);
 
 	void removeAllElements()
 	{
@@ -376,7 +376,7 @@ template <class Element> bool Vector<Element>::addElement(Element* objp)
 	return true;
 }
 
-template <class Element> bool Vector<Element>::ensureCapacity(unsigned int minCapacity)
+template <class Element> bool Vector<Element>::ensureCapacity(size_t minCapacity)
 {
 	if(_data.size >= minCapacity) {
 		return true;
